Move some .h stuff to .cc and fix error in gui-test
